Set up initial aiohttp app
This app structure is based on https://github.com/lsst-sqre/templatebot
and https://github.com/lsst-sqre/sqrbot-jr. Now that we're getting the
hang of deploying aiohttp apps in api.lsst.codes, we can start to look
towards consolidating common code in a library.

Includes the the Travis CI set up:

- Runs tests
- Deploys a docker image