Add Avro serializer for the render_ready topic
After we create a GitHub repository we need to be able to send a message
to the templatebot-render_ready topic so that templatebot itself can
render and push the first commit.

This adds the serializer. Note that the Avro schema isn't maintained in
this repo; templatebot "owns" it. So here we're just getting the latest
schema given the subject name.