Implement Travis activation for technotes
This happens in two steps:

1. Sync the user's Travis account. We now wait for the sync to be
   complete. This is necessary for activation.
2. Activate the repository on Travis CI.

A difficulty with Travis CI right now is that we're using both the .org
and .com endpoints, depending on the organization. This is why many of
the functions in the travisci.py module take the slug so that we can
internally determine the appropriate endpoint.

Eventually this should be done automatically (probe the API to determine
if .com is available for the org, and otherwise default to the old .org
endpoint?