Implement PRs with LTD secrets for Travis
This commit implements the task of encrypting secrets needed for the
LSST the Docs upload client, adding those secrets to the .travis.yml
file, committing and pushing that change to GitHub, and then opening a
Pull Request so that the user can incorporate those secrets into the
document.

This workflow uses ruamel.yaml to edit the YAML file so that comments
are preserved.