Fix empty submodule PR issue
The updated GitPython didn't work with our original implementation of
the lsst-texmf submodule PR service; the PR would be submitted but the
branch would be empty (i.e., equal to the main branch).

This updates the GitPython usage in several ways and now works.
I think GitPython introduced some API improvements that we're able to
take advantage of now.